State 통합 어느정도 진행 (Init 시 Size가 필요한데 initState에서 처리 못하는 경우 어떻게 처리해야 할 지 생각해야함)
